# Add filter selectivity estimation for auto search strategy