Currently, ```AutoStrategy``` of search query uses a cost-based planner which needs to compute the exact selectivity of the filter specified in the query. According to the result of JMC profiling, I figured out the bottleneck is calling the expensive ```union()``` operation of bitmaps. So, I propose an alternative way to estimate filter selectivity. 

In this patch, I simply assumed dimension values are independent to calculate approximate selectivity. As a result, the calculated selectivity may not be exact.
However, I think it's worthwhile because the planning cost of ```AutoStrategy``` will be reduced, thereby that strategy can be used in more cases which will be helpful for users. In addition, we can improve the selectivity estimation algorithm if it is needed.

Here are some numbers.

*master*
```
concise bitmap

Benchmark                                  (limit)  (numSegments)  (rowsPerSegment)  (schemaAndQuery)  Mode  Cnt       Score      Error  Units
SearchBenchmark.queryMultiQueryableIndex    750000              1            750000           basic.B  avgt   25  463864.923 ± 5935.667  us/op
SearchBenchmark.querySingleQueryableIndex   750000              1            750000           basic.B  avgt   25  471846.438 ± 5391.567  us/op

roaring bitmap

Benchmark                                  (limit)  (numSegments)  (rowsPerSegment)  (schemaAndQuery)  Mode  Cnt       Score       Error  Units
SearchBenchmark.queryMultiQueryableIndex    750000              1            750000           basic.B  avgt   25  803411.999 ± 10526.744  us/op
SearchBenchmark.querySingleQueryableIndex   750000              1            750000           basic.B  avgt   25  807268.072 ± 10779.393  us/op
```

*patch*
```
concise bitmap

Benchmark                                  (limit)  (numSegments)  (rowsPerSegment)  (schemaAndQuery)  Mode  Cnt       Score      Error  Units
SearchBenchmark.queryMultiQueryableIndex    750000              1            750000           basic.B  avgt   25  301427.773 ± 4794.298  us/op
SearchBenchmark.querySingleQueryableIndex   750000              1            750000           basic.B  avgt   25  285926.630 ± 2592.590  us/op

roaring bitmap

Benchmark                                  (limit)  (numSegments)  (rowsPerSegment)  (schemaAndQuery)  Mode  Cnt       Score      Error  Units
SearchBenchmark.queryMultiQueryableIndex    750000              1            750000           basic.B  avgt   25  464101.372 ± 5595.434  us/op
SearchBenchmark.querySingleQueryableIndex   750000              1            750000           basic.B  avgt   25  466196.807 ± 3428.533  us/op
```

Also, you can see the overhead of selectivity estimation in the following graphs.
These graphs show the results of ```SearchBenchmark``` with the ```basic.C``` query.
*- Concise bitmap*
![concise](https://cloud.githubusercontent.com/assets/2322288/21928982/5b42d288-d9d1-11e6-95d6-6f2b8b88f80e.png)

*- Roaring bitmap*
![roaring](https://cloud.githubusercontent.com/assets/2322288/21928994/713c54d8-d9d1-11e6-9064-0a067bcae464.png)

I think the overhead of selectivity estimation is acceptable. If others agree, I'll change the default search strategy to ```auto```.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/druid-io/druid/3848)
<!-- Reviewable:end -->