@jihoonson 2017-01-20
@gianm, I don't have exact numbers comparing the time taken by bitmap lookup, getting size, and bitmap union, but I think the above SearchQuery benchmark result shows the performance comparison of summing bitmap sizes and bitmap union. When just summing bitmap sizes instead of unioning them, the performance improvement ratios were about 165% and 173% for concise and roaring bitmap types, respectively. 

I think unioning bitmaps is the bottleneck quite obviously. Here is a JMC profile result of SearchBenchmark.

![2017-01-20 8 46 25](https://cloud.githubusercontent.com/assets/2322288/22130676/1c6c0848-def1-11e6-8892-fd7c603a32b3.png)

As you can see, just unioning bitmaps contributes about 6% to the total execution time of a search query. MinMaxPriorityQueue.poll() takes most of time, and I/O (via mmap) contributes only 20% of the total cost of bitmap unions.