@jihoonson 2017-01-23
@gianm, addressed comments. Thank you for your review.