@jihoonson 2017-01-24
@leventov, thanks for your review. I addressed comments.