@jihoonson 2017-01-25
@leventov thanks for your detailed review. I addressed comments.