@jihoonson 2017-01-25
The Travis failure is due to Kafka connection timeout and looks not related to this issue.