@jihoonson 2017-01-27
Fix test failures.